--- PETsARD/evaluator/anonymeter/__init__.py
@@ -1,9 +1,0 @@
-from PETsARD.evaluator.anonymeter.anonymeter import Anonymeter
-from PETsARD.evaluator.anonymeter.anonymeter_singlingout_univariate import Anonymeter_SinglingOut_Univariate
-from PETsARD.evaluator.anonymeter.anonymeter_linkability import Anonymeter_Linkability
-from PETsARD.evaluator.anonymeter.anonymeter_inference import Anonymeter_Inference
-
-
-__all__ = [
-    'Anonymeter'
-]
